Fixed messagepack limit error (desc)
The messagepack error is a error message since we are parsing way too much data trough the NUI.

There's an limit as to how much data you can transmit (at once).

This commit moves the vehicle props which  previously was parsed to NUI to a table with plate as keys and values being the vehicle props.
This way we wont reach any limit any time soon.

Also this has minor performance improvements on the server side since we no longer decode vehicles on the server!